LayoutTests/platform/ios-sim/svg/animations/animate-text-nested-transforms-expected.txt

SVG 1.1 dynamic animation tests

This is some text.
test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL rootSVGElement.getBBox().x should be close to 117. Was 115.880615234375.
FAIL rootSVGElement.getBBox().y should be close to -160. Was -162.96607971191406.
PASS rootSVGElement.getBBox().x is 367
PASS rootSVGElement.getBBox().y is 550
PASS successfullyParsed is true

TEST COMPLETE
